haha

+++ PartBQuestion1.java
@@ -101,7 +101,7 @@ public class PartBQuestion1 {
             System.out.println("Items Quantity : " + totalquantity);
             System.out.println("Membership : Yes");
             System.out.println("Membership Discount : 10%");
-            System.out.println("Total Purchased : "+ totalprice);
+            System.out.println("Total Purchased : "+ (float)totalprice);
             System.out.println("Rebates : " + totalrebate);
             totalgst = gst.count(totalprice);
             totaldiscounts = g.calc(totalgst);
